feat(slides): add new IonSlides module for Swiper 7
