+++ core/src/components/slides/IonicSwiper.ts
@@ -1,102 +1,128 @@
-export const IonicSwiper = (opts: any) => {
-  const { swiper, extendParams } = opts;
-  const slidesParams: any = {
-    effect: undefined,
-    direction: 'horizontal',
-    initialSlide: 0,
-    loop: false,
-    parallax: false,
-    slidesPerView: 1,
-    spaceBetween: 0,
-    speed: 300,
-    slidesPerColumn: 1,
-    slidesPerColumnFill: 'column',
-    slidesPerGroup: 1,
-    centeredSlides: false,
-    slidesOffsetBefore: 0,
-    slidesOffsetAfter: 0,
-    touchEventsTarget: 'container',
-    autoplay: false,
-    freeMode: false,
-    freeModeMomentum: true,
-    freeModeMomentumRatio: 1,
-    freeModeMomentumBounce: true,
-    freeModeMomentumBounceRatio: 1,
-    freeModeMomentumVelocityRatio: 1,
-    freeModeSticky: false,
-    freeModeMinimumVelocity: 0.02,
-    autoHeight: false,
-    setWrapperSize: false,
-    zoom: {
-      maxRatio: 3,
-      minRatio: 1,
-      toggle: false,
-    },
-    touchRatio: 1,
-    touchAngle: 45,
-    simulateTouch: true,
-    touchStartPreventDefault: false,
-    shortSwipes: true,
-    longSwipes: true,
-    longSwipesRatio: 0.5,
-    longSwipesMs: 300,
-    followFinger: true,
-    threshold: 0,
-    touchMoveStopPropagation: true,
-    touchReleaseOnEdges: false,
-    iOSEdgeSwipeDetection: false,
-    iOSEdgeSwipeThreshold: 20,
-    resistance: true,
-    resistanceRatio: 0.85,
-    watchSlidesProgress: false,
-    watchSlidesVisibility: false,
-    preventClicks: true,
-    preventClicksPropagation: true,
-    slideToClickedSlide: false,
-    loopAdditionalSlides: 0,
-    noSwiping: true,
-    runCallbacksOnInit: true,
-    coverflowEffect: {
-      rotate: 50,
-      stretch: 0,
-      depth: 100,
-      modifier: 1,
-      slideShadows: true
-    },
-    flipEffect: {
-      slideShadows: true,
-      limitRotation: true
-    },
-    cubeEffect: {
-      slideShadows: true,
-      shadow: true,
-      shadowOffset: 20,
-      shadowScale: 0.94
-    },
-    fadeEffect: {
-      crossFade: false
-    },
-    a11y: {
-      prevSlideMessage: 'Previous slide',
-      nextSlideMessage: 'Next slide',
-      firstSlideMessage: 'This is the first slide',
-      lastSlideMessage: 'This is the last slide'
+import { addEventListener, raf, removeEventListener } from '../../utils/helpers';
+
+/**
+ * This is a plugin for Swiper that allows it to work
+ * with Ionic Framework and the routing integrations.
+ * Without this plugin, Swiper would be incapable of correctly
+ * determining the dimensions of the slides component as
+ * each view is initially hidden before transitioning in.
+ */
+const setupSwiperInIonic = (swiper: any, watchForIonPageChanges = true) => {
+  if (typeof (window as any) === 'undefined') { return; }
+
+  const swiperEl = swiper.el;
+  const ionPage = swiperEl.closest('.ion-page');
+
+  if (!ionPage) {
+    if (watchForIonPageChanges) {
+
+      /**
+       * If no ion page found, it is possible
+       * that we are in the overlay setup step
+       * where the inner component has been
+       * created but not attached to the DOM yet.
+       * If so, wait for the .ion-page class to
+       * appear on the root div and re-run setup.
+       */
+      const rootNode = swiperEl.getRootNode();
+      if (rootNode.tagName === 'DIV') {
+        const mo = new MutationObserver((m: MutationRecord[]) => {
+          const mutation = m[0];
+          const wasEmpty = mutation.oldValue === null;
+          const hasIonPage = rootNode.classList.contains('ion-page');
+
+          /**
+           * Now that we have an .ion-page class
+           * we can safely attempt setup again.
+           */
+          if (wasEmpty && hasIonPage) {
+            mo.disconnect();
+
+            /**
+             * Set false here so we do not
+             * get infinite loops
+             */
+            setupSwiperInIonic(swiper, false);
+          }
+        });
+
+        mo.observe(rootNode, {
+          attributeFilter: ['class'],
+          attributeOldValue: true
+        });
+      }
     }
+    return;
   }
 
-  if (swiper.pagination) {
-    slidesParams.pagination = {
-      type: 'bullets',
-      clickable: false,
-      hideOnClick: false,
+  /**
+   * If using slides in a modal or
+   * popover we need to wait for the
+   * overlay to be shown as these components
+   * are hidden when they are initially created.
+   */
+  const modalOrPopover = swiperEl.closest('ion-modal, ion-popover');
+  if (modalOrPopover) {
+    const eventName = modalOrPopover.tagName === 'ION-MODAL' ? 'ionModalWillPresent' : 'ionPopoverWillPresent';
+    const overlayCallback = () => {
+      /**
+       * We need an raf here so the update
+       * is fired one tick after the overlay is shown.
+       */
+      raf(() => {
+        swiperEl.swiper.update();
+        removeEventListener(modalOrPopover, eventName, overlayCallback);
+      });
     }
+    addEventListener(modalOrPopover, eventName, overlayCallback);
+  } else {
+    /**
+     * If using slides in a page
+     * we need to wait for the ion-page-invisible
+     * class to be removed so Swiper can correctly
+     * compute the dimensions of the slides.
+     */
+    const mo = new MutationObserver((m: MutationRecord[]) => {
+      const mutation = m[0];
+      const wasPageHidden = mutation.oldValue?.includes('ion-page-invisible');
+      const isPageHidden = ionPage.classList.contains('ion-page-invisible');
+
+      /**
+       * Only update Swiper if the page was
+       * hidden but is no longer hidden.
+       */
+      if (!isPageHidden && isPageHidden !== wasPageHidden) {
+        swiperEl.swiper.update();
+      }
+    });
+
+    mo.observe(ionPage, {
+      attributeFilter: ['class'],
+      attributeOldValue: true
+    });
   }
 
-  if (swiper.scrollbar) {
-    slidesParams.scrollbar = {
-      hide: true
-    }
+  /**
+   * We also need to listen for the appload event
+   * which is emitted by Stencil in the
+   * event that Swiper is being used on the
+   * view that is rendered initially.
+   */
+  const onAppLoad = () => {
+    swiperEl.swiper.update();
+    removeEventListener(window, 'appload', onAppLoad);
   }
 
-  extendParams(slidesParams);
+  addEventListener(window, 'appload', onAppLoad);
+}
+
+export const IonicSwiper = {
+  name: 'ionic',
+  on: {
+    afterInit(swiper: any) {
+      console.warn('[Deprecation Warning]: The IonicSwiper module has been deprecated in favor of the IonSlides module. This change was made to better support the Swiper 7 release. The IonicSwiper module will be removed in Ionic 7.0. See https://ionicframework.com/docs/api/slides#migration for revised migration steps.');
+
+      setupSwiperInIonic(swiper);
+    }
+  }
 }
